backend-drf/requirements.txt
absl-py==2.3.1
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.13.4
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cryptography==45.0.4
curl_cffi==0.11.4
cycler==0.12.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flatbuffers==25.12.19
fonttools==4.58.5
frozendict==2.4.6
gast==0.7.0
google-pasta==0.2.0
grpcio==1.80.0
h5py==3.14.0
idna==3.10
joblib==1.5.1
keras==3.14.0
kiwisolver==1.4.8
libclang==18.1.1
markdown-it-py==3.0.0
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
multitasking==0.0.11
namex==0.1.0
numpy==2.3.1
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
protobuf==6.31.1
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.4
rich==14.0.0
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
wrapt==2.1.2
yfinance==0.2.65
